pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
